+++ src/wallet/service_test.go
@@ -131,7 +131,7 @@ func TestServiceCreateWallet(t *testing.T) {
 					Seed:     seed,
 					Encrypt:  tc.encrypt,
 					Password: tc.password,
-				})
+				}, nil)
 				require.Equal(t, tc.err, err)
 				if err != nil {
 					return
@@ -152,14 +152,14 @@ func TestServiceCreateWallet(t *testing.T) {
 				}
 
 				// create wallet with dup wallet name
-				_, err = s.CreateWallet(wltName, Options{Seed: "seed2"})
+				_, err = s.CreateWallet(wltName, Options{Seed: "seed2"}, nil)
 				require.Equal(t, err, ErrWalletNameConflict)
 
 				// create wallet with dup seed
 				dupWlt := "dup_wallet.wlt"
 				_, err = s.CreateWallet(dupWlt, Options{
 					Seed: seed,
-				})
+				}, nil)
 				require.EqualError(t, err, fmt.Sprintf("wallet %s would be duplicate with %v, same seed", dupWlt, wltName))
 
 				// check if the dup wallet is created
@@ -185,7 +185,6 @@ func TestServiceLoadWallet(t *testing.T) {
 	tt := []struct {
 		name          string
 		opts          Options
-		scanN         uint64
 		bg            BalanceGetter
 		err           error
 		expectAddrNum int
@@ -196,8 +195,8 @@ func TestServiceLoadWallet(t *testing.T) {
 			Options{
 				Seed:  "seed",
 				Label: "wallet",
+				ScanN: 5,
 			},
-			5,
 			mockBalanceGetter{
 				addrs[0]: BalancePair{Confirmed: Balance{Coins: 1e6, Hours: 100}},
 			},
@@ -210,8 +209,8 @@ func TestServiceLoadWallet(t *testing.T) {
 			Options{
 				Seed:  "seed",
 				Label: "wallet",
+				ScanN: 5,
 			},
-			5,
 			mockBalanceGetter{
 				addrs[1]: BalancePair{Confirmed: Balance{Coins: 1e6, Hours: 100}},
 			},
@@ -226,8 +225,8 @@ func TestServiceLoadWallet(t *testing.T) {
 				Label:    "wallet",
 				Encrypt:  true,
 				Password: []byte("pwd"),
+				ScanN:    5,
 			},
-			5,
 			mockBalanceGetter{
 				addrs[0]: BalancePair{Confirmed: Balance{Coins: 1e6, Hours: 100}},
 			},
@@ -242,8 +241,8 @@ func TestServiceLoadWallet(t *testing.T) {
 				Label:    "wallet",
 				Encrypt:  true,
 				Password: []byte("pwd"),
+				ScanN:    5,
 			},
-			5,
 			mockBalanceGetter{
 				addrs[1]: BalancePair{Confirmed: Balance{Coins: 1e6, Hours: 100}},
 			},
@@ -266,7 +265,7 @@ func TestServiceLoadWallet(t *testing.T) {
 				require.NoError(t, err)
 				wltName := newWalletFilename()
 
-				w, err := s.loadWallet(wltName, tc.opts, tc.scanN, tc.bg)
+				w, err := s.loadWallet(wltName, tc.opts, tc.bg)
 				require.Equal(t, tc.err, err)
 				if err != nil {
 					return
@@ -404,7 +403,7 @@ func TestServiceNewAddress(t *testing.T) {
 
 				wltName := newWalletFilename()
 
-				w, err := s.CreateWallet(wltName, tc.opts)
+				w, err := s.CreateWallet(wltName, tc.opts, nil)
 				if err != nil {
 					require.Equal(t, tc.expectErr, err)
 					return
@@ -576,7 +575,7 @@ func TestServiceGetWallets(t *testing.T) {
 				w2, err := s.CreateWallet(wltName, Options{
 					Seed:  "seed",
 					Label: "label",
-				})
+				}, nil)
 				require.NoError(t, err)
 				wallets = append(wallets, w2)
 
@@ -634,7 +633,7 @@ func TestServiceReloadWallets(t *testing.T) {
 			}
 
 			wltName := "t1.wlt"
-			w1, err := s.CreateWallet(wltName, Options{Seed: "seed1"})
+			w1, err := s.CreateWallet(wltName, Options{Seed: "seed1"}, nil)
 			require.NoError(t, err)
 
 			fmt.Println(dir, w1.Filename())
@@ -865,7 +864,7 @@ func TestServiceCreateAndSignTransaction(t *testing.T) {
 
 				wltName := newWalletFilename()
 
-				w, err := s.CreateWallet(wltName, tc.opts)
+				w, err := s.CreateWallet(wltName, tc.opts, nil)
 				require.NoError(t, err)
 
 				tx, err := s.CreateAndSignTransaction(w.Filename(), tc.pwd, tc.vld, unspents, uint64(headTime), tc.coins, tc.dest)
@@ -1734,7 +1733,7 @@ func TestServiceUpdateWalletLabel(t *testing.T) {
 				}
 
 				// Create a new wallet
-				w, err := s.CreateWallet(tc.wltName, tc.opts)
+				w, err := s.CreateWallet(tc.wltName, tc.opts, nil)
 				require.NoError(t, err)
 
 				err = s.UpdateWalletLabel(tc.updateWltName, tc.label)
@@ -1829,7 +1828,7 @@ func TestServiceEncryptWallet(t *testing.T) {
 				}
 
 				// Create a new wallet
-				w, err := s.CreateWallet(tc.wltName, tc.opts)
+				w, err := s.CreateWallet(tc.wltName, tc.opts, nil)
 				require.NoError(t, err)
 
 				// Encrypt the wallet
@@ -1957,7 +1956,7 @@ func TestServiceDecryptWallet(t *testing.T) {
 					return
 				}
 
-				_, err = s.CreateWallet(tc.wltName, tc.opts)
+				_, err = s.CreateWallet(tc.wltName, tc.opts, nil)
 				require.NoError(t, err)
 
 				_, err = s.DecryptWallet(tc.decryptWltName, tc.password)
@@ -2007,7 +2006,7 @@ func TestServiceDecryptWallet(t *testing.T) {
 	}
 }
 
-func TestServiceScanAheadWalletAddresses(t *testing.T) {
+func TestServiceCreateWalletWithScan(t *testing.T) {
 	bg := make(mockBalanceGetter, len(addrsOfSeed1))
 	addrs := fromAddrString(t, addrsOfSeed1)
 	for _, a := range addrs {
@@ -2026,8 +2025,6 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 	tt := []struct {
 		name             string
 		opts             Options
-		pwd              []byte
-		scanN            uint64
 		balGetter        BalanceGetter
 		disableWalletAPI bool
 		expect           exp
@@ -2054,7 +2051,6 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 				Encrypt:  true,
 				Password: []byte("pwd"),
 			},
-			pwd:       []byte("pwd"),
 			balGetter: bg,
 			expect: exp{
 				err:              nil,
@@ -2068,9 +2064,9 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 		{
 			name: "no coins and scan 1, unencrypted",
 			opts: Options{
-				Seed: "seed1",
+				Seed:  "seed1",
+				ScanN: 1,
 			},
-			scanN:     1,
 			balGetter: bg,
 			expect: exp{
 				err:              nil,
@@ -2087,9 +2083,8 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 				Seed:     "seed1",
 				Encrypt:  true,
 				Password: []byte("pwd"),
+				ScanN:    1,
 			},
-			pwd:       []byte("pwd"),
-			scanN:     1,
 			balGetter: bg,
 			expect: exp{
 				err:              nil,
@@ -2103,9 +2098,9 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 		{
 			name: "no coins and scan 10, unencrypted",
 			opts: Options{
-				Seed: "seed1",
+				Seed:  "seed1",
+				ScanN: 10,
 			},
-			scanN:     10,
 			balGetter: bg,
 			expect: exp{
 				err:              nil,
@@ -2119,9 +2114,9 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 		{
 			name: "scan 5 get 5, unencrypted",
 			opts: Options{
-				Seed: "seed1",
+				Seed:  "seed1",
+				ScanN: 5,
 			},
-			scanN: 5,
 			balGetter: mockBalanceGetter{
 				addrs[4]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
 			},
@@ -2140,9 +2135,8 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 				Seed:     "seed1",
 				Encrypt:  true,
 				Password: []byte("pwd"),
+				ScanN:    5,
 			},
-			pwd:   []byte("pwd"),
-			scanN: 5,
 			balGetter: mockBalanceGetter{
 				addrs[4]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
 			},
@@ -2158,9 +2152,9 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 		{
 			name: "scan 5 get 4, unencrypted",
 			opts: Options{
-				Seed: "seed1",
+				Seed:  "seed1",
+				ScanN: 5,
 			},
-			scanN: 5,
 			balGetter: mockBalanceGetter{
 				addrs[2]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
 				addrs[3]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
@@ -2180,9 +2174,8 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 				Seed:     "seed1",
 				Encrypt:  true,
 				Password: []byte("pwd"),
+				ScanN:    5,
 			},
-			pwd:   []byte("pwd"),
-			scanN: 5,
 			balGetter: mockBalanceGetter{
 				addrs[2]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
 				addrs[3]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
@@ -2199,9 +2192,9 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 		{
 			name: "scan 5 get 4 have 6, unencrypted",
 			opts: Options{
-				Seed: "seed1",
+				Seed:  "seed1",
+				ScanN: 5,
 			},
-			scanN: 5,
 			balGetter: mockBalanceGetter{
 				addrs[2]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
 				addrs[3]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
@@ -2219,9 +2212,9 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 		{
 			name: "confirmed and predicted, unencrypted",
 			opts: Options{
-				Seed: "seed1",
+				Seed:  "seed1",
+				ScanN: 5,
 			},
-			scanN: 5,
 			balGetter: mockBalanceGetter{
 				addrs[2]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
 				addrs[3]: BalancePair{Predicted: Balance{Coins: 10, Hours: 100}},
@@ -2241,9 +2234,8 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 				Seed:     "seed1",
 				Encrypt:  true,
 				Password: []byte("pwd"),
+				ScanN:    5,
 			},
-			pwd:   []byte("pwd"),
-			scanN: 5,
 			balGetter: mockBalanceGetter{
 				addrs[2]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
 				addrs[3]: BalancePair{Predicted: Balance{Coins: 10, Hours: 100}},
@@ -2258,31 +2250,13 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 			},
 		},
 		{
-			name: "confirmed and predicted, wrong password",
-			opts: Options{
-				Seed:     "seed1",
-				Encrypt:  true,
-				Password: []byte("pwd"),
-			},
-			pwd:   []byte("wrong password"),
-			scanN: 5,
-			balGetter: mockBalanceGetter{
-				addrs[2]: BalancePair{Confirmed: Balance{Coins: 10, Hours: 100}},
-				addrs[3]: BalancePair{Predicted: Balance{Coins: 10, Hours: 100}},
-			},
-			expect: exp{
-				err: ErrInvalidPassword,
-			},
-		},
-		{
 			name: "wallet api disabled",
 			opts: Options{
 				Seed:     "seed",
 				Encrypt:  true,
 				Password: []byte("pwd"),
+				ScanN:    5,
 			},
-			pwd:              []byte("pwd"),
-			scanN:            5,
 			balGetter:        mockBalanceGetter{},
 			disableWalletAPI: true,
 			expect: exp{
@@ -2303,27 +2277,17 @@ func TestServiceScanAheadWalletAddresses(t *testing.T) {
 				})
 				require.NoError(t, err)
 
-				if tc.disableWalletAPI {
-					_, err = s.ScanAheadWalletAddresses("", tc.pwd, tc.scanN, tc.balGetter)
-					require.Equal(t, tc.expect.err, err)
-					return
-				}
-
 				wltName := newWalletFilename()
-				w, err := s.CreateWallet(wltName, tc.opts)
-				require.NoError(t, err)
-
-				require.NoError(t, w.Validate())
-
-				w1, err := s.ScanAheadWalletAddresses(wltName, tc.pwd, tc.scanN, tc.balGetter)
+				w, err := s.CreateWallet(wltName, tc.opts, tc.balGetter)
 				require.Equal(t, tc.expect.err, err)
 				if err != nil {
 					return
 				}
 
-				require.Len(t, w1.Entries, tc.expect.entryNum)
-				for i := range w1.Entries {
-					require.Equal(t, addrsOfSeed1[i], w1.Entries[i].Address.String())
+				require.NoError(t, w.Validate())
+				require.Len(t, w.Entries, tc.expect.entryNum)
+				for i := range w.Entries {
+					require.Equal(t, addrsOfSeed1[i], w.Entries[i].Address.String())
 				}
 			})
 		}
@@ -2426,7 +2390,7 @@ func TestGetWalletSeed(t *testing.T) {
 				}
 
 				// Create a wallet
-				_, err = s.CreateWallet(tc.wltName, tc.opts)
+				_, err = s.CreateWallet(tc.wltName, tc.opts, nil)
 				require.NoError(t, err)
 
 				seed, err := s.GetWalletSeed(tc.id, tc.pwd)